requirements.txt
sounddevice==0.4.6
numpy==1.24.3
openai-whisper==20231117
google-generativeai==0.3.2
pyautogui==0.9.54
keyboard==0.13.5
python-dotenv==1.0.0
auto-py-to-exe==2.36.0
pydub==0.25.1
SpeechRecognition==3.10.0